Update traffic-mirroring-considerations.md
Re-wrote the sentence for clarity. Original example is very confusing and math shown is incorrect since It does not say which direction the mirrored traffic is being done (inbound only, outbound only or combined inbound + outbound). I used the combined as it is simple to understand and many of the VPC Traffic mirroring on-line tech talks in the past have also shown the same example as well. If you are not clear on the math, please see: https://youtu.be/zPovlZxuZ-c?t=1646